edk/vector/Matrix:  - Fix the bug where delete a matrix with error inside the createMatrix function.  - Change the typeTemplate to receive zero. Use the memset.  - Change the cloneFrom function to copy the full line instead value per value.  - Implement functions to add, sub, increment, decrement and compare if the matrix is equal to another matrix or if the size are equal.  - Implement the operators to the matrix. To use operators, I discomment the cantDelete function and the usage of canDelete variable.

+++ edk/vector/Matrix.h
@@ -47,16 +47,21 @@ template <class typeTemplate>
 class Matrix{
 public:
     Matrix(){
+        this->canDeleteVector=true;
+        this->matrixSize = 0u;
         //set the matrix to NULL
         this->matrix=NULL;
+        //set can delete vector
     }
     Matrix(edk::size2f32 size){
+        this->canDeleteVector=true;
         //set the matrix to NULL
         this->matrix=NULL;
         //create a new matrix
         this->createMatrix(size);
     }
     Matrix(edk::uint32 width,edk::uint32 height){
+        this->canDeleteVector=true;
         //set the matrix to NULL
         this->matrix=NULL;
         //create a new matrix
@@ -64,10 +69,12 @@ public:
     }
     ~Matrix(){
         //
-        //if(this->canDeleteVector){
+        if(this->canDeleteVector){
             this->deleteMatrix();
-        //}
-        this->canDeleteVector=false;
+        }
+        else{
+            this->canDeleteVector=true;
+        }
     }
 
 
@@ -87,7 +94,7 @@ public:
                         memset(this->matrix[i],0u,sizeof(typeTemplate)*size.width);
                     }
                     else{
-                        for(edk::uint32 j=i+1u;i>0u;i--){
+                        for(edk::uint32 j=i+1u;i>0u;j--){
                             //delete the matrix
                             delete[] this->matrix[j];
                         }
@@ -162,19 +169,213 @@ public:
     //return the value in a position
     typeTemplate get(edk::vec2ui32 position){
         //
-        typeTemplate ret;
         if(this->have(position)){
             //return the variable
             return this->matrix[position.y][position.x];
         }
-#pragma GCC diagnostic push
-#pragma GCC diagnostic ignored "-Wmaybe-uninitialized"
+        typeTemplate ret; memset((void*)&ret,0u,sizeof(typeTemplate));
         return ret;
-#pragma GCC diagnostic pop
     }
     //return the value in a position
     typeTemplate get(edk::uint32 x,edk::uint32 y){
         return this->get(edk::vec2ui32(x,y));
+    }
+
+    //test if a matrix is equal then other
+    bool isEqual(edk::vector::Matrix<typeTemplate>* matrix){
+        if(matrix){
+            if(this->matrix
+                    && matrix->matrix
+                    && this->matrixSize.width==matrix->matrixSize.width
+                    && this->matrixSize.height==matrix->matrixSize.height
+                    ){
+                for(edk::uint32 y=0u;y<this->matrixSize.height;y++){
+                    for(edk::uint32 x=0u;x<this->matrixSize.width;x++){
+                        if(this->matrix[y][x]!=matrix->matrix[y][x]){
+                            return false;
+                        }
+                    }
+                }
+                return true;
+            }
+        }
+        return false;
+    }
+    bool isEqualSize(edk::vector::Matrix<typeTemplate>* matrix){
+        if(matrix){
+            if(this->matrix
+                    && matrix->matrix
+                    && this->matrixSize.width==matrix->matrixSize.width
+                    && this->matrixSize.height==matrix->matrixSize.height
+                    ){
+                return true;
+            }
+        }
+        return false;
+    }
+    //add the matrix with another
+    bool add(edk::vector::Matrix<typeTemplate>* matrix){
+        //test if the matrix is equal in size
+        if(this->isEqualSize(matrix)){
+            for(edk::uint32 y=0u;y<this->matrixSize.height;y++){
+                for(edk::uint32 x=0u;x<this->matrixSize.width;x++){
+                    this->matrix[y][x]=this->matrix[y][x]+matrix->matrix[y][x];
+                }
+            }
+            return true;
+        }
+        return false;
+    }
+    //add the matrix with another
+    bool increment(edk::int32 n){
+        //test if the matrix is equal in size
+        if(this->matrix && this->matrixSize.width && this->matrixSize.height){
+            for(edk::uint32 y=0u;y<this->matrixSize.height;y++){
+                for(edk::uint32 x=0u;x<this->matrixSize.width;x++){
+                    this->matrix[y][x]=this->matrix[y][x]+n;
+                }
+            }
+            return true;
+        }
+        return false;
+    }
+    //sub the matrix with another
+    bool sub(edk::vector::Matrix<typeTemplate>* matrix){
+        //test if the matrix is equal in size
+        if(this->isEqualSize(matrix)){
+            for(edk::uint32 y=0u;y<this->matrixSize.height;y++){
+                for(edk::uint32 x=0u;x<this->matrixSize.width;x++){
+                    this->matrix[y][x]=this->matrix[y][x]-matrix->matrix[y][x];
+                }
+            }
+            return true;
+        }
+        return false;
+    }
+    //add the matrix with another
+    bool decrement(edk::int32 n){
+        //test if the matrix is equal in size
+        if(this->matrix && this->matrixSize.width && this->matrixSize.height){
+            for(edk::uint32 y=0u;y<this->matrixSize.height;y++){
+                for(edk::uint32 x=0u;x<this->matrixSize.width;x++){
+                    this->matrix[y][x]=this->matrix[y][x]-n;
+                }
+            }
+            return true;
+        }
+        return false;
+    }
+    bool multiply(edk::vector::Matrix<typeTemplate>* matrix){
+        if(this->matrix && this->matrixSize.width && this->matrixSize.height
+                &&
+                matrix->matrix && matrix->matrixSize.width && matrix->matrixSize.height
+                ){
+            //first compare if can have the product of the matrix
+            if(this->matrixSize.width == matrix->matrixSize.height){
+                //alloc the new matrix
+                edk::size2ui32 size = edk::size2ui32(matrix->matrixSize.width,this->matrixSize.height);
+
+
+                typeTemplate** newMatrix = new typeTemplate*[size.height];
+                typeTemplate value;
+                if(newMatrix){
+                    memset(newMatrix,0u,sizeof(typeTemplate)*size.height);
+                    for(edk::uint32 y=0u;y<size.height;y++){
+                        newMatrix[y] = new typeTemplate[size.width];
+                        if(newMatrix[y]){
+                            memset(newMatrix[y],0u,sizeof(typeTemplate)*size.width);
+                            for(edk::uint32 x=0u;x<size.width;x++){
+                                //clean the value
+                                memset((void*)&value,0u,sizeof(typeTemplate));
+                                //multiply
+                                for(edk::uint32 i=0u;i<this->matrixSize.width;i++){
+                                    value = value + (this->matrix[y][i] * matrix->matrix[i][x]);
+                                }
+                                //set the value
+                                newMatrix[y][x]=value;
+                            }
+                        }
+                        else{
+                            for(edk::uint32 j=y+1u;y>0u;j--){
+                                //delete the matrix
+                                delete[] newMatrix[j];
+                            }
+                            delete[] newMatrix;
+                            return false;
+                        }
+                    }
+                    //in the end. delete the old matrix and copy the new matrix
+                    this->deleteMatrix();
+                    this->matrix = newMatrix;
+                    this->matrixSize = size;
+                    return true;
+                }
+            }
+        }
+        return false;
+    }
+    bool multiply(edk::vector::Matrix<typeTemplate>* matrix1,edk::vector::Matrix<typeTemplate>* matrix2){
+        if(matrix1->matrix && matrix1->matrixSize.width && matrix1->matrixSize.height
+                &&
+                matrix2->matrix && matrix2->matrixSize.width && matrix2->matrixSize.height
+                ){
+            //first compare if can have the product of the matrix
+            if(matrix1->matrixSize.width == matrix2->matrixSize.height){
+                //alloc the new matrix
+                edk::size2ui32 size = edk::size2ui32(matrix2->matrixSize.width,matrix1->matrixSize.height);
+
+
+                typeTemplate** newMatrix = new typeTemplate*[size.height];
+                typeTemplate value;
+                if(newMatrix){
+                    memset(newMatrix,0u,sizeof(typeTemplate)*size.height);
+                    for(edk::uint32 y=0u;y<size.height;y++){
+                        newMatrix[y] = new typeTemplate[size.width];
+                        if(newMatrix[y]){
+                            memset(newMatrix[y],0u,sizeof(typeTemplate)*size.width);
+                            for(edk::uint32 x=0u;x<size.width;x++){
+                                //clean the value
+                                memset((void*)&value,0u,sizeof(typeTemplate));
+                                //multiply
+                                for(edk::uint32 i=0u;i<matrix1->matrixSize.width;i++){
+                                    value = value + (matrix1->matrix[y][i] * matrix2->matrix[i][x]);
+                                }
+                                //set the value
+                                newMatrix[y][x]=value;
+                            }
+                        }
+                        else{
+                            for(edk::uint32 j=y+1u;y>0u;j--){
+                                //delete the matrix
+                                delete[] newMatrix[j];
+                            }
+                            delete[] newMatrix;
+                            return false;
+                        }
+                    }
+                    //in the end. delete the old matrix and copy the new matrix
+                    this->deleteMatrix();
+                    this->matrix = newMatrix;
+                    this->matrixSize = size;
+                    return true;
+                }
+            }
+        }
+        return false;
+    }
+
+    //print the matrix
+    bool printMatrix(){
+        if(this->matrix && this->matrixSize.width && this->matrixSize.height){
+            printf("\nMatrix [%u] [%u]",this->matrixSize.width,this->matrixSize.height);
+            for(edk::uint32 y=0u;y<this->matrixSize.height;y++){
+                for(edk::uint32 x=0u;x<this->matrixSize.width;x++){
+                    this->printElement(x,y,&this->matrix[y][x]);
+                }
+            }fflush(stdout);
+            return true;
+        }
+        return false;
     }
 
     //delete the array
@@ -192,13 +393,6 @@ public:
         this->canDeleteVector=false;
     }
 
-
-    //cant
-    void cantDeleteVector(){
-        //
-        this->canDeleteVector=false;
-    }
-
     bool cloneFrom(edk::vector::Matrix<typeTemplate>* matrix){
         this->deleteMatrix();
         if(matrix){
@@ -206,9 +400,7 @@ public:
                 if(this->createMatrix(matrix->matrixSize.width,matrix->matrixSize.height)){
                     //copy the values
                     for(edk::uint32 y=0u;y<matrix->matrixSize.height;y++){
-                        for(edk::uint32 x=0u;x<matrix->matrixSize.width;x++){
-                            this->matrix[y][x] = matrix->matrix[y][x];
-                        }
+                        memcpy(this->matrix[y],matrix->matrix[y],sizeof(typeTemplate)*matrix->matrixSize.width);
                     }
                     return true;
                 }
@@ -217,12 +409,74 @@ public:
         return false;
     }
 
+    edk::vector::Matrix<typeTemplate>* operator=(edk::vector::Matrix<typeTemplate> matrix){
+        this->cloneFrom(&matrix);
+        matrix.cantDeleteVector();
+        return this;
+    }
+    bool operator==(edk::vector::Matrix<typeTemplate> matrix){
+        matrix.cantDeleteVector();
+        return this->isEqual(&matrix);
+    }
+    bool operator!=(edk::vector::Matrix<typeTemplate> matrix){
+        matrix.cantDeleteVector();
+        return !this->isEqual(&matrix);
+    }
+    edk::vector::Matrix<typeTemplate> operator+(edk::vector::Matrix<typeTemplate> matrix){
+        this->add(&matrix);
+        return matrix;
+    }
+    edk::vector::Matrix<typeTemplate> operator++(edk::int32 n){
+        this->increment(n);
+        return matrix;
+    }
+    edk::vector::Matrix<typeTemplate> operator++(){
+        this->increment(1);
+        return matrix;
+    }
+    void operator+=(edk::vector::Matrix<typeTemplate> matrix){
+        this->add(&matrix);
+    }
+    edk::vector::Matrix<typeTemplate> operator-(edk::vector::Matrix<typeTemplate> matrix){
+        this->sub(&matrix);
+        return matrix;
+    }
+    edk::vector::Matrix<typeTemplate> operator--(){
+        this->decrement(1);
+        return matrix;
+    }
+    edk::vector::Matrix<typeTemplate> operator--(edk::int32 n){
+        this->decrement(n);
+        return matrix;
+    }
+    void operator-=(edk::vector::Matrix<typeTemplate> matrix){
+        this->sub(&matrix);
+    }
+    edk::vector::Matrix<typeTemplate> operator*(edk::vector::Matrix<typeTemplate> matrix){
+        edk::vector::Matrix<typeTemplate> ret;
+        ret.multiply(this,&matrix);
+        return ret;
+    }
+    edk::vector::Matrix<typeTemplate> operator*=(edk::vector::Matrix<typeTemplate> matrix){
+        this->multiply(&matrix);
+        return *this;
+    }
+protected:
+    void printElement(edk::uint32 x,edk::uint32 y,typeTemplate* value){
+        printf("\n[%u][%u] == [%ld]",x,y,(edk::int64)*value);
+    }
+
 private:
     typeTemplate** matrix;
     //size of the vector
     edk::size2ui32 matrixSize;
     //test if can delete the vector
     bool canDeleteVector;
+
+    //cant
+    void cantDeleteVector(){
+        this->canDeleteVector=false;
+    }
 };
 }//end namespace vector
 }//end namespace edk
